feat(target-profile): liquid compiler target profile system with R-SPU formal verification
- Implement TargetConfig in ast/program.rs: dynamic word_bits, reg_bits,
  guard_bits decoupled from hardcoded 32-bit assumptions
- Parser (module_parser) now reads target {} blocks and propagates config
  through the full pipeline (width solver, emitters, toolchain)
- All emitters (verilog, rspu, riscv, arm, sexpr, rspu_encoding) updated
  to respect target profile bit-widths dynamically
- Width solver bounded at 64-bit ceiling with E504/E505 diagnostics for
  overflow detection (warn-and-continue for Yosys truncation)
- HLS optimizer, MAPE-K bridge, and temporal emitter adapted to new arch
- rspu_top 16-core design formally verified: BMC depth 10, Z3 solver,
  all SVA assertions PASSED (28 min, 894 signals, 483 reflexes)
- New target_profile_tests.rs integration test suite
- Test suite updated to accommodate 64-bit architecture transition

+++ src/emit/rspu_encoding/opcodes.rs
@@ -53,13 +53,13 @@ pub const OP_TAG_BRANCH: u8 = 37;
 /// Total number of assigned opcodes (used + reserved).
 pub const TOTAL_OPCODES: usize = 38;
 
-/// Maximum value for a 10-bit immediate field.
+/// Maximum value for a 10-bit immediate field (used in AluImm).
 pub(super) const IMM10_MAX: u64 = 0x3FF;
 
 // ---------------------------------------------------------------------------
 // Encoded instruction newtype
 // ---------------------------------------------------------------------------
 
-/// A 32-bit encoded R-SPU instruction word.
+/// A 64-bit encoded R-SPU instruction word (R-SPU 2.0).
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-pub struct EncodedInstruction(pub u32);
+pub struct EncodedInstruction(pub u64);
